+++ examples/poem/tokio-metrics/src/main.rs
@@ -1,0 +1,26 @@
+use poem::middleware::TokioMetrics;
+use poem::{
+    get, handler, listener::TcpListener, middleware::Tracing, web::Path, EndpointExt, Route, Server,
+};
+
+#[handler]
+fn hello(Path(name): Path<String>) -> String {
+    format!("hello: {}", name)
+}
+
+#[tokio::main]
+async fn main() -> Result<(), std::io::Error> {
+    if std::env::var_os("RUST_LOG").is_none() {
+        std::env::set_var("RUST_LOG", "poem=debug");
+    }
+    tracing_subscriber::fmt::init();
+
+    let app = Route::new()
+        .at("/hello/:name", get(hello))
+        .with(Tracing)
+        .with(TokioMetrics::new("/metrics"));
+    Server::new(TcpListener::bind("127.0.0.1:3000"))
+        .name("hello-world")
+        .run(app)
+        .await
+}