+++ poem/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.16] 2022-3-18
+
+- Add `Cache-Control: no-cache` header to SSE response.
+- Add `TokioMetrics` middleware. [#206](https://github.com/poem-web/poem/issues/206)
+
 # [1.3.15] 2022-3-16
 
 - Check the `Content-type` when use `Json` extractor. [#236](https://github.com/poem-web/poem/pull/236)